新增StoreManagerService
