+++ PlusOneBuy/FBPlusOneBuy/Controllers/SettingController.cs
@@ -27,7 +27,7 @@ namespace FBPlusOneBuy.Controllers
             var fpage_repo = new FanPagesRepository();
             var fpage = fpage_repo.SelectBinding(userid);
 
-            int managerId = LineBindingService.GetManagerId(userid);
+            int managerId = StoreManagerService.GetManagerId(userid);
             var groupList= LineBindingService.GetGroupList(managerId);
             ViewBag.groupList = groupList;
             ViewBag.bindingPage = fpage;
